Compare fix and complete
